+++ src/components/ScreenWrapper.tsx
@@ -1,0 +1,63 @@
+import React from 'react';
+import { View, Text, StyleSheet, SafeAreaView, ScrollView } from 'react-native';
+import { theme } from '../styles/theme';
+
+interface ScreenWrapperProps {
+  title?: string;
+  children: React.ReactNode;
+  style?: any;
+  contentStyle?: any;
+  scrollViewStyle?: any;
+}
+
+export const ScreenWrapper: React.FC<ScreenWrapperProps> = ({
+  title,
+  children,
+  style,
+  contentStyle,
+  scrollViewStyle,
+}) => {
+  return (
+    <SafeAreaView style={[styles.container, style]}>
+      {title && (
+        <View style={styles.header}>
+          <Text style={styles.title}>{title}</Text>
+        </View>
+      )}
+      <ScrollView
+        style={[styles.scrollView, scrollViewStyle]}
+        contentContainerStyle={[styles.contentContainer, contentStyle]}
+        showsVerticalScrollIndicator={false}
+      >
+        {children}
+      </ScrollView>
+    </SafeAreaView>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: theme.colors.background,
+  },
+  header: {
+    paddingHorizontal: 20,
+    paddingVertical: 16,
+    borderBottomWidth: 1,
+    borderBottomColor: theme.colors.border,
+    backgroundColor: theme.colors.surface,
+  },
+  title: {
+    fontSize: theme.typography.sizes.xl,
+    fontWeight: theme.typography.weights.bold,
+    color: theme.colors.primary,
+    fontFamily: theme.typography.fontFamily,
+  },
+  scrollView: {
+    flex: 1,
+  },
+  contentContainer: {
+    paddingHorizontal: 20,
+    paddingVertical: 20,
+  },
+});
